Sanitize input for negative numbers and non-number variables
I created two helper functions: errorIfNegative, which returns an error
essage when a negative number is fed as an argument; and errorIfNotNumber,
which throws an error when there's a non-number object fed as an
argument. Both of these wrte created in a TDD method, which is taking some
getting used to but I  will try to get better at as time goes on.