Add errorWhenInputIsFloat
I added functionality to my validator that returns an error when a user
enters a value that has a decimal point in it. For now, the program is
not going to include functionality for floats or fractions,
though that is something I would like to add in a future update.