Add tests for validator and validator-helpers
I continued working on the validator function.
I am running in to a lot of trouble, and I fear I am trying too hard to
predict how user input it going to vary. Going forward, I am going to
start working on the algorithms themselves and work on a validator when
I set up my front end so I have a better idea of the input I will be
receiving from users.