Finish multiplier tool
I finished my multiplier tool, which takes two numbers and a method of
multiplication and returns the product of those two numbers. It utilizes
several helper functions that convert the numbers to an array, then feed
that array to the designated multiplication algorithm. It then returns
the product from that computation.
One improvement is that is now allows for negative multiplication, by
using a boolean value to determine if the product should be negative or
not.