Implement traditional multiplication algorithm
I added functionality for the traditional multiplication algorithm.
Using string conversion, I am able to add 0's to the end of single-digit
multiplication to simulate numbers being added in the tens/hundreds/etc.
place. This resembless how humans use think about multiplication, which
is in line with the goal of the program: to determine how much faster
certain multiplication methods are compared to others.