Finish validator.js so it passes all unit tests
I tweaked the order of the code for the validator function to ensure it
passes all unit tests. There were issues with the test accepting
different types of values, and the order the validator was calling the
helper function was messing up error calls.
In the future, I would like to create some form of error handler so I
can easily keep track of error messages and trace errors to their
source.